Help text to use the build.py script and extensions to .gitignore

+++ shared_lib_gen_consume_example/build.py
@@ -47,26 +47,32 @@ allOpts = {
                     }
           }
 
-if detectPlatform == "Linux":
-    print("Detected platform - LINUX")
-    opt = allOpts["linux"][sys.argv[1]]
-    seperator = " ; "
-elif detectPlatform == "Windows":
-    print("Detected platform - Windows")
-    opt = allOpts["win"][sys.argv[1]]
-    seperator = " & " 
+if len(sys.argv) <= 1:
+    print("Usage:")
+    print("python3 build.py <project>")
+    print("Example: python3 build.py calc --> Builds the calc project")
+    print("Example: python3 build.py mymath --> Builds the mymath project")
+else:
+    if detectPlatform == "Linux":
+        print("Detected platform - LINUX")
+        opt = allOpts["linux"][sys.argv[1]]
+        seperator = " ; "
+    elif detectPlatform == "Windows":
+        print("Detected platform - Windows")
+        opt = allOpts["win"][sys.argv[1]]
+        seperator = " & " 
 
-print("Removing specified directories")
-for d in opt["delDirs"]:
-    shutil.rmtree(d, ignore_errors=True)
+    print("Removing specified directories")
+    for d in opt["delDirs"]:
+        shutil.rmtree(d, ignore_errors=True)
 
-print("Creating empty build directory:", opt["createDirs"])
-for d in opt["createDirs"]:
-    os.mkdir(d)
+    print("Creating empty build directory:", opt["createDirs"])
+    for d in opt["createDirs"]:
+        os.mkdir(d)
 
-print("Running cmake, building and installing")
-finalCmd = "cd " + opt["workingDir"] + seperator + opt["cmakeCmd"] + seperator + opt["build"] + seperator + opt["install"]
-subprocess.Popen(finalCmd, shell=True) 
+    print("Running cmake, building and installing")
+    finalCmd = "cd " + opt["workingDir"] + seperator + opt["cmakeCmd"] + seperator + opt["build"] + seperator + opt["install"]
+    subprocess.Popen(finalCmd, shell=True)
